// src/features/proforma-details/data/schema.ts
import { z } from 'zod'

export const proformaDetailSchema = z.object({
  id: z.number(),
  proformaId: z.number(),
  productServiceId: z.number().nullable(),
  description: z.string(),
  unit: z.string().nullable(),
  quantity: z.number(),
  unitPrice: z.number(),
  subtotal: z.number(),
  tax: z.number().nullable(),
  total: z.number(),
  order: z.number().nullable(),
  createdAt: z.string(),
  updatedAt: z.string(),
})

export type ProformaDetail = z.infer<typeof proformaDetailSchema>